Setup / Mobile App v_1_0_0 development : * **Moved:** All test files were moved from the single __tests__ directory at the root of the project to instead live inside a __tests__ directory that is located in the same directory as the file being tested. * **Moved:** All mock files for keeping them closer to their associated files. * **Added:** Eslint was added as well as a configuration file for configuring eslint rules for the project. * **Added:** New testing suite was created for testing the DeviceStorageService with an empty test.
