+++ SFCC-RN-Components/src/lib/OCAPIService/__mocks__/CategoryMock.js
