+++ SFCC-RN-Components/src/services/deviceStorage/__tests__/DeviceStorageServiceTests.js
@@ -1,0 +1,19 @@
+/**
+ * @file DeviceStorageServiceTests.js
+ * @fileoverview - Provides unit test coverage of the DeviceStorageService class
+ * and all of its critical public member methods.
+ */
+
+describe('src/services/deviceStorage/DeviceStorageService.js -- Tests', () => {
+
+  // Encryption Method Tests
+  describe('DeviceStorageService.encryptValues() -- Tests', () => {
+    // Suite level variable declaration.
+    test('should encrypt data for storage on a device.', () => {
+      expect(true);
+    });
+  });
+
+});
+
+
