+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "SFCC-React",
+  "name": "sfcc-react",
   "version": "0.1.0",
   "description": "POC for use of ReactJS with Sales Force Commerce Cloud e-commerce platform.",
   "main": "index.js",
@@ -10,6 +10,7 @@
   "author": "Galen Goforth <galengoforth@gmail.com>",
   "license": "MIT",
   "dependencies": {
+    "jest": "^22.4.2",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-native-cli": "^2.0.1"
